backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-offer-insight-backend"
version = "0.1.0"
description = "FastAPI backend for Job Offer Insight"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "sqlmodel==0.0.24",
    "alembic==1.16.5",
    "pyyaml==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
